pyproject.toml
[project]
name = "atlas"
version = "0.1.0"
description = "ATLAS — Assistente Pessoal com Personalidade"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "openai>=1.60",
    "chromadb>=0.6",
    "tavily-python>=0.5",
    "google-api-python-client>=2.160",
    "google-auth-oauthlib>=1.2",
    "pydantic-settings>=2.7",
    "python-frontmatter>=1.1",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "numpy>=1.26",
    "httpx>=0.28",
    "edge-tts>=6.1",
    "duckduckgo-search>=6.0",
    "groq>=0.4",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
